+++ package.json
@@ -56,6 +56,7 @@
     "@types/react": "^18.2.65",
     "@types/webpack": "^5.28.5",
     "babel-loader": "^9.1.3",
+    "circular-dependency-plugin": "^5.2.2",
     "formiojs": "~4.13.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^3.2.5",
