+++ src/components/FormioComponent.tsx
@@ -2,6 +2,12 @@ import type {AnyComponentSchema} from '@open-formulieren/types';
 
 import {getRegistryEntry} from '@/registry';
 
+/**
+ * The props that every field render component must support.
+ *
+ * The interface/props are generic in the component schema - make sure to narrow this
+ * for the component-type specific props/component.
+ */
 export interface FormioComponentProps<S extends AnyComponentSchema = AnyComponentSchema> {
   /**
    * The Formio.js component definition, limited to the features supported in Open Forms.
@@ -21,7 +27,12 @@ const FormioComponent: React.FC<FormioComponentProps> = ({componentDefinition}) 
       </div>
     );
   }
-  return <TypeSpecificComponent componentDefinition={componentDefinition} />;
+  return (
+    <TypeSpecificComponent
+      componentDefinition={componentDefinition}
+      renderNested={FormioComponent}
+    />
+  );
 };
 
 export default FormioComponent;
