+++ src/registry/fieldset/index.tsx
@@ -1,12 +1,16 @@
 import type {FieldsetComponentSchema} from '@open-formulieren/types';
 
-import FormioComponent from '@/components/FormioComponent';
+import type {FormioComponentProps} from '@/components/FormioComponent';
 
 export interface FieldsetProps {
   componentDefinition: FieldsetComponentSchema;
+  renderNested: React.FC<FormioComponentProps>;
 }
 
-const Fieldset: React.FC<FieldsetProps> = ({componentDefinition: {components}}) => {
+const Fieldset: React.FC<FieldsetProps> = ({
+  componentDefinition: {components},
+  renderNested: FormioComponent,
+}) => {
   return (
     <div>
       I am a fieldset field with {components.length} child components. The nested children are:
